watchlist_simdump
Ever wanted to create a watchlist, but worried about how many hits it
would generate?

I created this script to help deal with that concern.

You can feed it a tab-delimited text file with console-style process
search queries in the "watchlist" column, and it will go count up how
many hits that query generated.

It also prints out what it found in tab-delimited format, so if you get
a noisy query, you can break it down in a pivot table and figure out how
to tune it.  You can tell it what fields you want it to output and
aggregate on, so it prints those field plus a count of all process hits
where those fields were exactly the same string, plus the name of the
watch list being tested.

Options include...

•"All fields", or specify a list of fields.
•Ensure no hostnames and/or sensor ID's are in the output.
•How many days of history to look through, by default 10.
•A bunch of throttles, including total max rows of output, max rows per
each tested watchlist, and "chunk sizes" on both returns from queries
and how many to aggregate before flushing to output (to protect against
memory consumption).  Note that the smaller you make aggregation size
before flush, the less efficient/consolidated the aggregation is.
•Secure anonymizer for hostnames and sensor ID's.
•If you specify "start" or "last_update", by how much to "truncate or
round" the timestamp, so as to maximize aggregation.