cb_print_tab_output
Are you the sort that likes to export a list of stuff and pop it into
Excel?  This script is designed to make it easy to do that, with at
least a few things you can get from the CB Client API.

Specifically you can export all the items in the "results" part of the
JSON that comes from a query to /alert, /binary, /process, or /sensor.
The script accepts your query syntax, which it will pass to the "?q="
parameter of the GET request..

Or another way to think of it is, if you know how to type raw query
syntax into the console UI for a process, alert, sensor or binary
search, you can get this script to export those results.

Output goes to STDOUT in tab-delimited format.

Parameters include CB Server name, your API token, the query, the node (
process, alert, sensor, binary ), the fields you want in the output ( or
"--allfields"), and optionally maximum number of results to output, how
many results to process per iteration of the GET loop (to help control
script memory consumption), sort order, a field to sort on, and whether
to verify the SSL connection.

Here's an example command line:

./cb_print_tab_output.py -t <removed> -s cb.demo.local -q 'start:-240h'
-f "process_name,start,last_update,filemod_count" -n process -u start -o
asc > results.txt

This will get 10 days worth of data from /process, sort on start time in
ascending order, and print only the short list of fields mentioned.