move_sensors_csv_input
is file will take a CSV file as input and use the data to move sensors
from one group to another.  The format of the file requires a "Computer"
and a "Group" column.  In the "Computer" column you put the sensor's
regular name, and in the "Group" name you put the name of the sensor
group.  You specify the API token, server name, and file name on the
command line.  The script also has options to force or ignore SSL
certificate verification, and to print some verbose logging to standard
out.