Move Computers Around Bulk
This VBScript uses the Bit9 Admin API (not the RESTful API) to move a
bunch of computers from one policy to another.  It has a few different
options, including using an input file, using a source policy, and
limiting how many machines are moved in one run.  The Bit9 Admin API is
not installed on the Bit9 Server by default.  If you do not already have
it, contact Technical Support.