Computer Tag Bulk Update
This VBScript uses the Bit9 Admin API (not the RESTful API) to set the
"Computer Tag" field on a bunch of computers.  It has a lot of options,
including using queries, using input files, doing fuzzy matching, and so
on.  The Bit9 Admin API is not installed by default on the Bit9 Server,
so if you do not already have it, contact Technical Support.